apprequirements.txt
Flask-RESTful
Flask-SQLAlchemy
# Flask_Jwt
Flask-JWT-Extened
 